+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 * Update HTML Page Title on Data export Pages. Refs UIDEXP-369
 * Time Reported on the Progress Display. Refs UIDEXP-373.
 * Display error messages instead of error codes in Data export. Refs UIDEXP-371.
+* Suppressing fields on export - integrate with BE. Refs UIDEXP-375.
 
 ## [6.1.0](https://github.com/folio-org/ui-data-export/tree/v6.1.0) (2024-03-19)